[#34] Seperate all the endpoints (Manga and Chapters)
