+++ manga/urls.py
@@ -4,13 +4,13 @@ from . import views
 
 
 urlpatterns = [
-    path("", views.MangaCreateListApiView.as_view()),
-    # path("test/", views.MangaGenericListCreateApiView.as_view()),
+    path("", views.MangaGenericListCreateApiView.as_view()),
 
-    path("<str:manga>/", views.MangaRetrieveUpdateDestroy.as_view()),
-    # path("<str:manga>/test/", views.MangaGenericListCreateApiView.as_view()),
+    path("<str:id_name>/", views.MangaGenericsRetrieveUpdateDestroy.as_view()),
 
-    path("<str:manga>/chapters/", views.ChapterCreateListApiView.as_view()),
-    path("<str:manga>/chapters/<str:number>/",
-         views.ChapterRetreiveUpdateDestroy.as_view()),
+    # path("<str:id_name>/chapters/",
+    #      views.ChapterGenericListCreateApiView.as_view()),
+
+    # path("<str:id_name>/chapters/<str:name>/",
+    #      views.ChapterGenericsRetreiveUpdateDestroy.as_view()),
 ]
