+++ omiku/settings.py
@@ -42,6 +42,7 @@ INSTALLED_APPS = [
     'rest_framework_swagger',
     "rest_framework.authtoken",
     "manga",
+    "chapters",
 ]
 
 MIDDLEWARE = [
